The initial version grew buckets 1 item at a time. This resulted in the lowest memory allocation, at the cost of creating a lot of garbage (thus slowing inserts).
Using append() performed better by generating far less intermediary objects,
but could result in worse overall memory usage.

This approach strikes a balance, a vector is used to limit the number of
garbage, but its growth is linear (default 8).

The exact implications depends on how full the set is and how random the keys
are. But on a few cases, this gives almost perfect final size (initial case)
without as much garbage (and thus better insert performance).